+++ src/mania/score_state.rs
@@ -1,35 +1,32 @@
-/// Aggregation for a score's current state i.e. what are the current hitresults.
-#[derive(Clone, Debug, Default, Eq, PartialEq)]
+/// Aggregation for a score's current state.
+#[derive(Clone, Debug, Default, PartialEq, Eq)]
 pub struct ManiaScoreState {
     /// Amount of current 320s.
-    pub n320: usize,
+    pub n320: u32,
     /// Amount of current 300s.
-    pub n300: usize,
+    pub n300: u32,
     /// Amount of current 200s.
-    pub n200: usize,
+    pub n200: u32,
     /// Amount of current 100s.
-    pub n100: usize,
+    pub n100: u32,
     /// Amount of current 50s.
-    pub n50: usize,
+    pub n50: u32,
     /// Amount of current misses.
-    pub n_misses: usize,
+    pub n_misses: u32,
 }
 
 impl ManiaScoreState {
     /// Create a new empty score state.
-    #[inline]
     pub fn new() -> Self {
         Self::default()
     }
 
     /// Return the total amount of hits by adding everything up.
-    #[inline]
-    pub const fn total_hits(&self) -> usize {
+    pub const fn total_hits(&self) -> u32 {
         self.n320 + self.n300 + self.n200 + self.n100 + self.n50 + self.n_misses
     }
 
     /// Calculate the accuracy between `0.0` and `1.0` for this state.
-    #[inline]
     pub fn accuracy(&self) -> f64 {
         let total_hits = self.total_hits();
 
@@ -40,6 +37,6 @@ impl ManiaScoreState {
         let numerator = 6 * (self.n320 + self.n300) + 4 * self.n200 + 2 * self.n100 + self.n50;
         let denominator = 6 * total_hits;
 
-        numerator as f64 / denominator as f64
+        f64::from(numerator) / f64::from(denominator)
     }
 }
